Create a random number distribution.
The user picks a random number generator seed, an array size, and the maximum value a number in the array can be. The output is a list of the random numbers, the count of each number between 1 and the provided number by the user, the random distribution statistics. The user can use the trackbar to look up the value at a certain index in the array.